[refactor] change class name & remove unused Appear Animation
- 안쓰는 애니메이션 제거 및 이름 재설정
